// src/types/forms.ts
export interface FormValues {
    firstName: string;
    lastName: string;
    documentType: string;
    documentNumber: string;
    email: string;
    phoneNumber: string;
    file: File[];
    useSameData: boolean;
    billingFirstName: string;
    billingLastName: string;
    billingDocumentType: string;
    billingDocumentNumber: string;
    billingEmail: string;
    billingPhoneNumber: string;
  }
  